Fixed longest chain abort condition
The implementation was flawed and estimated the longest potential chain
far too generously. This commit fixes the estimation algorithm to be
slightly more conservative (and gains massive performance!). In the
future, we might be able to get even lower estimates for potential chain
length! We just have to take care that we never estimate TOO low.